Fixed the AR + new movement system
The AR system is now fixed
Detects the floor and the walls 
The debug plane prefab is now transparent, so you can't see the detection, but I've tried to crush on the floor a several times, it was successful.
New movement system - very easy to control the player.
fixed collision box - was under the player and not on it.
fixed player starting above the camera problem - now it's starting in front of it (the bug derived from the method: placecharacter()).